backend/requirements.txt
fastapi==0.100.1
uvicorn==0.23.1
networkx==3.1
pydantic==2.1.1
python-multipart==0.0.6
deepdiff==6.3.1
python-arango==7.8.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
